Eliminated CDNs and changed page format
I switched from CDNs to Bower package manager. Reformatted the page so that the control panel is at the top of the screen so that the adjustments in the slider will be more accurate. Also made a space for a drop down menu for when I add the marker color/class feature.